support for default values in json representation
to improve performance, protobuf 3 does not put default values on the wire.  to improve usability, they can be optionally added to the json representation.  this change makes that feature visible.  it is expected to be useful for metrics code that wants a complete record that includes all values, see https://app.zenhub.com/workspace/o/adsabs/adsmasterpipeline/issues/22